+++ src/DataPipeline.jl
@@ -37,7 +37,7 @@ include("core.jl")
 
 include("api.jl")
 export initialise, finalise
-export link_read!, link_write!
+export link_read!, link_read_files!, link_write!
 export read_array, read_table, read_distribution, read_estimate
 export write_array, write_table, write_distribution, write_estimate
 export raise_issue